ai/requirements.txt
uvicorn~=0.34.2
fastapi~=0.115.12
protobuf~=3.20.3
pydantic~=2.11.4
pip~=25.1
google-genai~=1.16.1